Refactor: 상품 등록, 수정 페이지 통합
Refs: #7
